stable slash version with permissions

+++ kandora.bot/models/LeagueConfig.cs
@@ -15,23 +15,23 @@ namespace kandora.bot.models
 
         public int Id { get; set; }
         public bool CountPoints { get; set; }
-        public float StartingPoints { get; set; }
+        public double StartingPoints { get; set; }
         public bool AllowSanma { get; set; }
-        public float Uma3p1 { get; set; }
-        public float Uma3p2 { get; set; }
-        public float Uma3p3 { get; set; }
-        public float Uma4p1 { get; set; }
-        public float Uma4p2 { get; set; }
-        public float Uma4p3 { get; set; }
-        public float Uma4p4 { get; set; }
-        public float Oka { get; set; }
-        public float PenaltyLast { get; set; }
+        public double Uma3p1 { get; set; }
+        public double Uma3p2 { get; set; }
+        public double Uma3p3 { get; set; }
+        public double Uma4p1 { get; set; }
+        public double Uma4p2 { get; set; }
+        public double Uma4p3 { get; set; }
+        public double Uma4p4 { get; set; }
+        public double Oka { get; set; }
+        public double PenaltyLast { get; set; }
         public bool UseEloSystem { get; set; }
-        public float InitialElo { get; set; }
-        public float MinElo { get; set; }
-        public float BaseEloChangeDampening { get; set; }
-        public float EloChangeStartRatio { get; set; }
-        public float EloChangeEndRatio { get; set; }
+        public double InitialElo { get; set; }
+        public double MinElo { get; set; }
+        public double BaseEloChangeDampening { get; set; }
+        public double EloChangeStartRatio { get; set; }
+        public double EloChangeEndRatio { get; set; }
         public int TrialPeriodDuration { get; set; }
 
     }
